feat: Implement notification module with CRUD operations and push notifications
- Added NotificationController for handling notification-related requests.
- Created NotificationResponse, UnreadCountResponse, RegisterDeviceRequest, and UnregisterDeviceRequest DTOs.
- Developed Notification and DeviceToken entities with JPA annotations.
- Implemented NotificationService and its implementation for managing notifications and device tokens.
- Created NotificationEventListener to handle events and send notifications asynchronously.
- Added repositories for Notification and DeviceToken entities.
- Created SQL migration scripts for notifications and device tokens tables.
- Implemented unit tests for ReviewServiceImpl and PayoutServiceImpl to ensure functionality.

+++ src/main/java/com/unishare/api/modules/mentor/dto/MentorResponse.java
@@ -16,7 +16,7 @@ public class MentorResponse {
     private String headline;
     private String expertise;
     private BigDecimal basePrice;
-    private Float ratingAvg;
+    private Double ratingAvg;
     private Integer sessionsCompleted;
     private String verificationStatus;
 }
